Using the Serialization interface, whatever there is of it, to load in PAK files. Using the pak interface, it's not at all documented, which needs to change, and there needs to be a dedicated interface to loading these files.
A lot of this is hold over from grad school and it's awful, just awful

+++ Source/Musa/Animation/Clip.cpp
@@ -13,7 +13,7 @@ AnimationClip::AnimationClip(const tchar* animationName, KeyFrame* frames, uint3
 	frameCount(count),
 	frameRate(frameRate_)
 {
-	uint32 len = Strlen(animationName);
+	size_t len = Strlen(animationName);
 	if (len >= ClipNameLength)
 	{
 		len = ClipNameLength - 1;
